backend/requirements.txt
Flask==0.12
pyjwt==1.4.2
flask-cors==3.0.2
google-auth==0.5.0
requests==2.12.4
requests-toolbelt==0.7.0
flask-restful
